watch-together/backend/requirements.txt
fastapi
python-socketio[asyncio_server]
uvicorn[standard]
python-dotenv
